requirements.txt
streamlit
google-generativeai
streamlit-mic-recorder
gtts
requests